Disable the floor_log2 inline optimization
I was not able to get this to work while building the cross-compiler with GCC
10.3.1.  It seems the problem is C++ rules are used for the linker even though
toplev.c is compiled with C rules.  The extra definition causes an error for
the C++ linker.

It may be better to solve this using 'static inline' instead of 'extern
inline', but I'm not sure.